+++ processor/src/test/fixtures/bad_input/com/example/fragment/includes/PrimitiveInIncludesModel.java
@@ -1,0 +1,8 @@
+package com.example.fragment.includes;
+
+import sting.Fragment;
+
+@Fragment( includes = int.class )
+public interface PrimitiveInIncludesModel
+{
+}