+++ processor/src/test/fixtures/bad_input/com/example/injector/includes/PrimitiveInIncludesModel.java
@@ -1,0 +1,15 @@
+package com.example.injector.includes;
+
+import sting.Injectable;
+import sting.Injector;
+
+@Injector( includes = byte.class )
+public interface PrimitiveInIncludesModel
+{
+  MyModel getMyModel();
+
+  @Injectable
+  class MyModel
+  {
+  }
+}