+++ processor/src/test/java/sting/processor/StingProcessorTest.java
@@ -446,6 +446,8 @@ public final class StingProcessorTest
                       "@Fragment target has an 'includes' parameter containing the value com.example.fragment.includes.MultipleProvidersIncludesModel.MyComponent that is annotated by multiple @StingProvider annotations. Matching annotations:\n" +
                       "    com.example.fragment.includes.MultipleProvidersIncludesModel.MyFrameworkComponent1\n" +
                       "    com.example.fragment.includes.MultipleProvidersIncludesModel.MyFrameworkComponent2" },
+        new Object[]{ "com.example.fragment.includes.PrimitiveInIncludesModel",
+                      "@Fragment target must not include a primitive in the includes parameter" },
         new Object[]{ "com.example.fragment.includes.UnannotatedProviderIncludesModel",
                       "@Fragment target has an parameter named 'includes' containing the value com.example.fragment.includes.UnannotatedProviderIncludesModel.MyComponent and that type is annotated by the @StingProvider annotation. The provider annotation expects a provider class named com.example.fragment.includes.UnannotatedProviderIncludesModel.MyComponent_Provider but that class is not annotated with either @Injector or @Fragment" },
 
@@ -650,6 +652,8 @@ public final class StingProcessorTest
                       "@Injector target has an 'includes' parameter containing the value com.example.injector.includes.MultipleProvidersIncludesModel.MyComponent that is annotated by multiple @StingProvider annotations. Matching annotations:\n" +
                       "    com.example.injector.includes.MultipleProvidersIncludesModel.MyFrameworkComponent1\n" +
                       "    com.example.injector.includes.MultipleProvidersIncludesModel.MyFrameworkComponent2" },
+        new Object[]{ "com.example.injector.includes.PrimitiveInIncludesModel",
+                      "@Injector target must not include a primitive in the includes parameter" },
         new Object[]{ "com.example.injector.includes.UnannotatedProviderIncludesModel",
                       "@Injector target has an parameter named 'includes' containing the value com.example.injector.includes.UnannotatedProviderIncludesModel.MyComponent and that type is annotated by the @StingProvider annotation. The provider annotation expects a provider class named com.example.injector.includes.UnannotatedProviderIncludesModel.MyComponent_Provider but that class is not annotated with either @Injector or @Fragment" },